Pretty much done, could use some better code styling to reduce argument passing

+++ lsbi/rc4.c
@@ -21,7 +21,7 @@ RC4(const uint8_t *msg, const uint8_t *bmpFile, const size_t msgSize)
     uint8_t *keyStream    = PRGA(s, msgSize);
     uint8_t *encryptedMsg = malloc(sizeof(*encryptedMsg) * msgSize);
 
-    for (int i = 0; i < msgSize; i++)
+    for (size_t i = 0; i < msgSize; i++)
         encryptedMsg[i] = msg[i] ^ keyStream[i];
 
     free(key);
@@ -63,7 +63,7 @@ initVectorS()
 {
     uint8_t *s = malloc(sizeof(*s) *   VECTOR_SIZE);
 
-    for (int i = 0; i < VECTOR_SIZE; i++)
+    for (size_t i = 0; i < VECTOR_SIZE; i++)
         s[i] = (uint8_t) i;
 
     return s;
@@ -74,7 +74,7 @@ initVectorT(const uint8_t *key)
 {
     uint8_t *t = malloc(sizeof(*t) * VECTOR_SIZE);
 
-    for (int i = 0; i < VECTOR_SIZE; i++)
+    for (size_t i = 0; i < VECTOR_SIZE; i++)
         t[i] = key[i % KEY_SIZE];
 
     return t;
